+++ back-end/src/main/java/com/tdd/backend/user/controller/UserController.java
@@ -4,10 +4,12 @@ import javax.validation.Valid;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.tdd.backend.auth.data.JwtTokenPairResponse;
@@ -47,5 +49,11 @@ public class UserController {
 	@PostMapping("/login")
 	public ResponseEntity<JwtTokenPairResponse> login(@RequestBody @Valid UserLogin userLogin) {
 		return ResponseEntity.ok(userService.login(userLogin));
+	}
+
+	@Operation(summary = "유저 로그아웃 요청", description = "사용자가 로그아웃을 하면 저장소에서 Refresh Token을 삭제하여 사용이 불가능하도록 한다.")
+	@DeleteMapping("/logout")
+	public void logout(@RequestHeader("Authorization") String refreshToken) {
+		userService.logout(refreshToken);
 	}
 }
