+++ back-end/src/test/java/com/tdd/backend/user/UserServiceTest.java
@@ -12,8 +12,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.tdd.backend.auth.jwt.RefreshTokenStorage;
 import com.tdd.backend.auth.encrypt.EncryptHelper;
+import com.tdd.backend.auth.jwt.RefreshTokenService;
 import com.tdd.backend.user.data.User;
 import com.tdd.backend.user.data.UserCreate;
 import com.tdd.backend.user.data.UserLogin;
@@ -33,6 +33,9 @@ class UserServiceTest {
 
 	@Autowired
 	private EncryptHelper encryptHelper;
+
+	@Autowired
+	RefreshTokenService refreshTokenService;
 
 	@Test
 	@DisplayName("유저 저장")
@@ -78,7 +81,7 @@ class UserServiceTest {
 		User findUser = userRepository.findByEmail("test@test.com").orElseThrow(UserNotFoundException::new);
 		softAssertions.assertThat(findUser.getId()).isEqualTo(user.getId());
 
-		softAssertions.assertThat(RefreshTokenStorage.isValidateUserId(user.getId())).isTrue();
+		softAssertions.assertThat(refreshTokenService.isRefreshTokenExists(user.getId())).isTrue();
 		softAssertions.assertAll();
 	}
